Pages refactor
