+++ src/lib/components/PostCard.svelte
@@ -8,19 +8,18 @@
 </script>
 
 <div
-  class="card"
+  class="card flex flex-col justify-center"
   on:click={() => {
     goto(link);
   }}
 >
-  <h1 class="font-bold text-xl">{data.title}</h1>
+  <h1 class="font-bold text-2xl">{data.title}</h1>
   <h2>{data.desc}</h2>
   <h2>{data.date}</h2>
 </div>
 
 <style lang="postcss">
   .card {
-    @apply hover:scale-102.5 rounded-lg shadow-lg m-4 p-4 h-56;
-    @apply cursor-pointer rounded-lg shadow-lg transition-transform;
+    @apply expand-sm rounded-lg shadow-lg m-4 p-4 h-56 cursor-pointer;
   }
 </style>
