+++ src/lib/components/ext.ts
@@ -1,0 +1,4 @@
+export default {
+  target: "_blank",
+  rel: "noopener",
+};
