+++ src/lib/data/repos.json
@@ -40,7 +40,7 @@
         {
             "title": "My Portfolio",
             "link": "https://github.com/Leomotors/portfolio-sv",
-            "desc": "This Website you are seeing right now! Written in SvelteKit + TailwindCSS and no UI Framework",
+            "desc": "THIS Website you are seeing right now! Written in SvelteKit + TailwindCSS and no UI Framework",
             "language": "Svelte",
             "img": "https://img.freepik.com/free-vector/blue-pink-halftone-background_53876-99004.jpg?size=626&ext=jpg"
         },
