+++ src/posts/220215.md
@@ -1,12 +1,17 @@
 # Parser Testing
 
-~~This is my first blog post~~
+My first blog post that is not about first blog post.
 
-Testing `marked`, `highlight.js`, My Vite Plugin and `@tailwindcss/typography`
+> My `real` first blog post coming soon!
 
-Styled with @tailwindcss/typography
+Here what you are seeing is me currently testing this blog system.
 
-Let me sleep first, I will look into this in the future
+The posts are stored in markdown files `*.md` preprocessed by my custom Vite Plugin
+at compile time which is located in `svelte.config.cjs`, during the preprocess,
+`marked` and `highlight.js` is used. After preprocess, the blog posts are now ready.
+
+Then it is injected to DOM using Svelte's `@html` directive. And styles with
+`@tailwindcss/typography`. The block code is styled by `highlight.js`'s css.
 
 ## this is h2
 
@@ -58,3 +63,7 @@ const a = `${bruh} yes`
 | 4   | 🌟🌟  |
 | 5   | 🌟🌟  |
 | 6 |  |
+
+## bruh
+
+an really long sentences Lorem ipsum dolor sit, amet consectetur adipisicing elit. Iure a harum libero consequuntur culpa eveniet accusantium laudantium quia, quod quos perspiciatis, quas commodi modi eos, ab quis. Mollitia odit assumenda molestias praesentium maxime fuga ut temporibus vero voluptatibus quisquam ipsum facilis quam officia aliquid cum, iste, error quae ab. Excepturi, ab? Reprehenderit, praesentium laboriosam? Suscipit quam molestiae tempore expedita vitae officia minima nesciunt, omnis recusandae assumenda autem repellendus libero optio, repudiandae voluptates harum ducimus nostrum odio, a quia illo sunt eaque. Quibusdam molestiae, deserunt eaque eligendi reiciendis dolorum quo, est eius dolor nemo voluptatem animi minus repellendus. Deserunt, consequuntur reiciendis!
