+++ src/routes/about.svelte
@@ -1,2 +1,14 @@
-<h1 class="page-title">Under Construction!</h1>
-<h2 class="page-subtitle">Website undergoing refactor</h2>
+<script lang="ts">
+  import LangCards from "$lib/sections/LangCards.svelte";
+  import EdHistory from "$lib/sections/EdHistory.svelte";
+  import ActivitiesCard from "$lib/sections/ActivitiesCard.svelte";
+  import ProjectsCard from "$lib/sections/ProjectsCard.svelte";
+</script>
+
+<main class="page-content">
+  <h1 class="page-title">About me</h1>
+  <LangCards />
+  <EdHistory />
+  <ActivitiesCard />
+  <ProjectsCard />
+</main>
